requirements.txt
django>=5.2.7
djangorestframework
python-dotenv
dj-database-url
psycopg2-binary
PyJWT

# If your DB uses bcrypt password hashes, also add:
# bcrypt
